✨: tighten mobile form width
