+++ frontend/src/pages/docs/md/changelog/20251101.md
@@ -123,6 +123,8 @@ What's DSPACE, you ask? You must be new around here, and if so, welcome! I'm gla
 
 -   Added Playwright coverage to verify the mobile "More" menu toggle hides and reveals unpinned
     navigation links when tapped.
+-   Hardened the custom item creation page for phones by tightening its padding and promoting the
+    mobile form Playwright check from backlog into the main test suite.
 
 -   [x] **Developer Ergonomics & Infrastructure Debt** 💯 (multi‑PR initiative)
     -   [x] **Unify package management (pnpm workspaces) 💯**
