+++ frontend/src/pages/inventory/create.astro
@@ -43,4 +43,10 @@ const title = "Create Custom Item";
     .form-container {
         margin-top: 2rem;
     }
-</style> 
+
+    @media (max-width: 600px) {
+        main {
+            padding: 1rem;
+        }
+    }
+</style>
